Initial commit of the strange attractor code.
This is pretty simple.  The program reads an input file listing
parameters describing the fractal (some samples are provided in the
examples/ directory).  These parameters can be overridden with
command-line arguments for making animations.  See the README for more
information.

Many of the examples are taken directly from Paul Bourke's website:
http://local.wasp.uwa.edu.au/~pbourke/fractals/